dashboard/requirements.txt
streamlit
pandas
plotly
psycopg2-binary 